requirements.txt
pygame-ce
particlepy
numpy
